LayoutTests/platform/ios-sim-deprecated/storage/indexeddb/intversion-persistence-expected.txt

CONSOLE MESSAGE: line 23: TypeError: undefined is not an object (evaluating 'request.onerror = unexpectedErrorCallback')
Test that integer versions are retained when backing store is opened/closed/re-opened.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

dbname = "intversion-persistence.html"
dbname = dbname + Date.now()

openFirstTime():
indexedDB.open(dbname, 1)
FAIL indexedDB.open(dbname, 1) threw exception TypeError: null is not an object (evaluating 'indexedDB.open')
FAIL Unexpected error: TypeError: undefined is not an object (evaluating 'request.onerror = unexpectedErrorCallback')
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
